Vendor `is-supported-regexp-flag` for CommonJS version
Include a vendored version of the `is-supported-regexp-flag` package
converted to CommonJS. For completeness, the original license is
included in the file but is otherwise omitted as we still depend on the
original package and hence cover its license transitively for consumers.
Similarly, despite it being vendored, by depending on the original
package we will become aware of any security advisories for the vendored
code through the `npm audit` command.