Add CommonJS support (#5)
Add support for usage by CommonJS modules as a dependency. This was
done because we want to be able to support users that do want to make
use of the non-backtracking regular expression engine but are not yet on
a Node.js version that supports importing ESM modules from CommonJS.

Includes a vendored version of the `is-supported-regexp-flag` package
converted to CommonJS. For completeness, the original license is
included in the file but is otherwise omitted as we still depend on the
original package and hence cover its license transitively for consumers.
Similarly, despite it being vendored, by depending on the original
package we will become aware of any security advisories for the vendored
code through the `npm audit` command.

This commit also includes a version bump so that this change can be
published to the npm registry.