Set up testing (#9)
Create a test suite for this package itself as well as vendored code.

The own tests ensures that the experimental linear-time engine is used
when the `--enable-experimental-regexp-engine` option is used (and
sanity checks the test when it is not used) - identical for ESM and CJS.

The vendored code test suite is simply a mechanistic conversion of
https://github.com/sindresorhus/is-supported-regexp-flag/blob/719e931/test.js

Furthermore, `.npmignore` ensures only relevant files are included
when publishing to the npm registry and `Justfile` offers standardized
commands for running the tests.