Set up benchmarking (#8)
Create a benchmark test file targeting the performance of using the
library's custom RegExp constructor.

The goal of this benchmark is to be able to identify performance
degradations w.r.t. language alternative and ideally reduce them. This
commit does not actually aim to do that, instead just provides the test
infrastructure.