Specify and test all supported versions
Add `package.json#engines` to list the supported versions of Node.js.
Anything above v15 is supported because the Node.js runtime supports the
experimental (linear-time) regular expression eninge. v14.18.0 and above
is supported because the present consumers of this library require that.
v12.20.0 and above is supported because `is-supported-regexp-flag`
supports it anyway, so we might as well. Instead of using `>=12.20.0` we
conservatively specify all supported releases explicitly to avoid
promising support for feature Node.js releases that we don't or can't
support.

To ensure compatibility, CI has been updated to run the tests on all
versions declared in `package.json#engines`.