Specify and test all supported versions (#16)
Add `package.json#engines` to list the supported versions of Node.js.
Anything above v15 is supported because the Node.js runtime supports the
experimental (linear-time) regular expression engine. v14.18.0 and above
is supported because the present consumers of this library require that.
Additionally, `"^12.20.0 || ^14.13.0` is supported because it is in
range for `is-supported-regexp-flag` -  so we might as well - and the
current test suite just works. We conservatively specify all supported
releases explicitly (rather than, say `>=14.13.0`) to avoid promising
support for feature Node.js releases that we don't (or can't) support.
To ensure compatibility, CI has been updated to run the tests on all
versions declared in `package.json#engines`.

Some refactoring was necessary. In particular, we now import
`perf_hooks` instead of `node:perf_hooks` in the tests because otherwise
we get "Error: Cannot find module 'node:perf_hooks'" on older Node.js
versions (e.g. v15). The lockfile has been downgraded to v2 (from v3) in
order to be able to install dependencies on Node.js v12.20.0 and
v14.13.0.

Additionally, when running tests on Node.js prior to v15 the experiment
regular expression engine is not enabled (that is, the Node.js CLI flag
`--enable-experimental-regexp-engine` is not used) because the flag did
not yet exists, thus causing an error when invoking Node.js with it.

Lastly, based on having to repeatedly run the tests, some flakiness in
the test was uncovered. Hence, the length of the string being matched
has been increased to avoid flakiness when not using the experiment
regexp engine. Now the duration with the experimental engine is far
below 1 (typically <0.1) while with it is far above 1 (typically >10).